feat: unified Environments page — merge environments + credentials
Consolidate the separate Environments (admin-only) and Credentials pages
into a single tabbed "Environments" page accessible to all users.

Backend:
- environment.list now open to all users (requireAuth instead of requireAdmin),
  enriched with hasUserKubeconfig/userConfigUpdatedAt, filtered by testOnly
- environment.update invalidates mismatched kubeconfigs when apiServer changes
- environment.delete pushes updated credentials to all affected AgentBoxes
- Added listFullForEnv() to UserEnvConfigRepository

Frontend:
- Unified tabbed page: Environments tab + Credentials tab
- Role-aware actions: all users upload/remove kubeconfig, admins edit/delete env
- Status badges: Configured (green) / Default (teal) / — (gray)
- Extracted EnvironmentDialog, KubeconfigUploadDialog, CredentialsSection components
- Sidebar: merged Credentials + Environments into single "Environments" entry
- /credentials route redirects to /environments?tab=credentials
